+++ src/GeoGen/random/RandomSequence2D.hpp
@@ -19,7 +19,7 @@ namespace geogen
 			inline void Advance() { this->current = this->mtRand(); }
 
 			inline int GetInt(Point point) const { return Hash(this->current, point.GetX(), point.GetY()); }
-			inline int GetInt(Point point, int min, int max) { return min + (this->GetInt(point) % (int)(max - min + 1)); }
+			inline int GetInt(Point point, int min, int max) { return min + ((unsigned)this->GetInt(point) % (int)(max - min + 1)); }
 			inline Height GetHeight(Point point) { return (Height)this->GetInt(point, HEIGHT_MIN, HEIGHT_MAX); }
 			inline double GetDouble(Point point) { return (double)(this->GetInt(point)) * (1. / 4294967295.); }
 		};
